Create copy-file.vbs
This script will copy a specified file to somewhere else, checking if the file already exists in the destination folder. If the file does exist in the destination folder it adds the date to the file-name, if it finds the file still already exists with the date, it adds more time-detail until it's gone down to seconds.

It does expect a .doc file, but it's a simple find-and-replace to change the extension.